templates/python-template/main.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
Script Principal de Automação
Author: Seu Nome
Date: 2026
Description: Descrição do que este script faz
"""

import os
from dotenv import load_dotenv
from utils import setup_logger

# Carrega variáveis de ambiente
load_dotenv()

# Configura logging
logger = setup_logger(__name__)


def main():
    """
    Função principal do script de automação.
    """
    logger.info("Iniciando automação...")
    
    try:
        # Sua lógica de automação aqui
        logger.info("Executando tarefa...")
        
        # Exemplo de uso de variável de ambiente
        debug_mode = os.getenv('DEBUG', 'False') == 'True'
        if debug_mode:
            logger.debug("Modo debug ativado")
        
        # Adicione sua lógica aqui
        
        logger.info("Automação concluída com sucesso!")
        
    except Exception as e:
        logger.error(f"Erro durante a execução: {str(e)}")
        raise


if __name__ == "__main__":
    main()
